git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@529 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/test/java/cspfj/priorityqueues/BinaryHeapTest.java
@@ -43,7 +43,7 @@ public final class BinaryHeapTest {
                 new Key<IdInteger>() {
 
                     @Override
-                    public double getKey(final IdInteger object) {
+                    public float getKey(final IdInteger object) {
                         return object.value;
                     }
                 });
