+++ src/test/java/cspfj/priorityqueues/FifoTest.java
@@ -42,7 +42,7 @@ public final class FifoTest {
         final Fifo<IdInteger> fifo = new Fifo<IdInteger>(new Key<IdInteger>() {
 
             @Override
-            public double getKey(final IdInteger object) {
+            public float getKey(final IdInteger object) {
                 return object.value;
             }
         });
