+++ src/test/java/cspfj/priorityqueues/SoftHeapTest.java
@@ -29,7 +29,7 @@ public class SoftHeapTest {
                 new Key<IdInteger>() {
 
                     @Override
-                    public double getKey(final IdInteger object) {
+                    public float getKey(final IdInteger object) {
                         return object.value;
                     }
